% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_string.R
\name{format_string}
\alias{format_string}
\title{Nicely format a string to title case}
\usage{
format_string(string, keep_caps = NULL)
}
\arguments{
\item{string}{The string to format.}

\item{keep_caps}{String patterns within the main string to make sure and keep capitalized
(e.g., abbreviations).}
}
\value{
A formatted character string.
}
\description{
Formats a string to title case while keeping some strings like roman numerals
and user specified patterns capitalized. Also ignores standard words that are
usually not capitalized in a title (e.g., and, or, the, in, etc.).
}
