% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longtail_barplot.R
\name{longtail_barplot}
\alias{longtail_barplot}
\title{Plot a longtail (non-normalized stacked) barplot}
\usage{
longtail_barplot(
  df,
  groups,
  column,
  negative_value,
  ylab = NULL,
  xlab = NULL,
  legendlab = NULL,
  color_list = NULL,
  flip_plot = FALSE,
  keep_caps = NULL,
  save = FALSE,
  figwidth = 1000,
  figheight = 1000,
  out_path = "composition_barplot.jpg"
)
}
\arguments{
\item{df}{Dataframe containing the grouping variable and categorical variable of interest.
Should be in a "long" format where each row is an observation (e.g., what
you get when reshaping a data.frame using \code{tidyr::pivot_longer()},
\code{reshape2::melt()}, \code{reshape()}, etc.).}

\item{groups}{Name of the categorical variable in \code{df} to group by (i.e., what each bar of
the plot will represent).}

\item{column}{Name of the categorical variable in \code{df} to be plotted (i.e., how each bar of
the plot will be broken up).}

\item{negative_value}{Name of the value in \code{column} that represents a "negative" or "absent" result.
(e.g., \code{0} for a binary column or \code{NULL}). Can be an actual value or \code{NA}.}

\item{ylab}{Title for the y-axis. (default is to use \verb{Frequency (\%)})}

\item{xlab}{Title for the x-axis. (default is to use name given to \code{groups})}

\item{legendlab}{Title for the legend. (default is to use name given to \code{column})}

\item{color_list}{Vector of R recognized color strings the length of the number of groups in
the variable provided to \code{column}.}

\item{flip_plot}{Whether to flip the plot so bars are now horizontal. (default: FALSE)}

\item{keep_caps}{Vector of character strings to make sure to keep capitalized. The function
automatically tries to keep roman numerals capitalized, but any other string
needs to be provided here.}

\item{save}{Whether to save the image to file. (default: FALSE)}

\item{figwidth}{Width of the output image file in pixels (\code{px}).}

\item{figheight}{Height of the output image file in pixels (\code{px}).}

\item{out_path}{File path for the output image file. Existing files at this path are overwritten.
Whatever extension included in the image file name (e.g., jpg, png, pdf) will
be the format that is outputted.}
}
\value{
A \code{ggplot2} figure object. If save is \code{TRUE}, also exports the image to file.
}
\description{
This function creates a non-normalized stacked bar plot reporting frequencies of
a categorical variable grouped by a specified grouping variable including total
frequencies for each group of the grouping variable.
}
