% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_barplot.R
\name{stratified_barplot}
\alias{stratified_barplot}
\title{Plot a Stratified Barplot Using Plotly}
\usage{
stratified_barplot(
  data,
  var,
  strata,
  ylab = NULL,
  xlab = NULL,
  legendlab = NULL,
  colors = NULL,
  test = c("fisher.test", "chisq.test"),
  alpha = 0.05
)
}
\arguments{
\item{data}{The data frame containing the variables of interest.}

\item{var}{The name of the categorical variable in \code{data} to be plotted.}

\item{strata}{The name of the categorical variable in \code{data} to group by.}

\item{ylab}{The title for the y-axis (default is to use \verb{Frequency (\%)}).}

\item{xlab}{The title for the x-axis (default is to use name given to \code{strata}).}

\item{legendlab}{The title for the legend (default is to use name given to \code{var}).}

\item{colors}{A vector of R recognized color strings the length of the number of groups
in the variable provided to \code{var} (vector).}

\item{test}{Which test to use for group comparison: \code{fisher.test} or \code{chisq.test}
(default: \code{fisher.test}).}

\item{alpha}{P-value threshold for significance (default: 0.05).}
}
\value{
A \code{plotly} figure object.
}
\description{
This function creates a bar plot reporting frequencies of a categorical variable
grouped by a specified strata variable. It can also test for significant differences
between groups using Fisher's exact test or Chi-squared test, annotating significant
comparisons in the plot.
}
