% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_barplot.R
\name{composition_barplot}
\alias{composition_barplot}
\title{Plot a compositional (normalized stacked) barplot}
\usage{
composition_barplot(
  df,
  groups,
  column,
  subrows = NULL,
  subcolumns = NULL,
  ylab = NULL,
  xlab = NULL,
  legendlab = NULL,
  color_list = NULL,
  remove_xaxis_text = FALSE,
  flip_plot = FALSE,
  save = FALSE,
  figwidth = 1000,
  figheight = 1000,
  out_path = "composition_barplot.jpg"
)
}
\arguments{
\item{df}{Dataframe containing the grouping variable and categorical variable of interest.
Should be in a "long" format where each row is an observation (e.g., what
you get when reshaping a data.frame using \code{tidyr::pivot_longer()},
\code{reshape2::melt()}, \code{reshape()}, etc.).}

\item{groups}{Name of the categorical variable in \code{df} to group by (i.e., what each bar of
the plot will represent).}

\item{column}{Name of the categorical variable in \code{df} to be plotted (i.e., how each bar of
the plot will be broken up).}

\item{subrows}{Name of a categorical variable to optionally stratify the plot by.}

\item{subcolumns}{Name of a categorical variable to optionally group bars by.}

\item{ylab}{Title for the y-axis. (default is to use \verb{Frequency (\%)})}

\item{xlab}{Title for the x-axis. (default is to use name given to \code{groups})}

\item{legendlab}{Title for the legend. (default is to use name given to \code{column})}

\item{color_list}{Vector of R recognized color strings the length of the number of groups in
the variable provided to \code{column}.}

\item{remove_xaxis_text}{Whether to remove the x-axis text and tick marks. Useful for when there are
many bars being plotted and showing the x-axis text is not feasible.
(default: FALSE)}

\item{flip_plot}{Whether to flip the plot so bars are now horizontal. (default: FALSE)}

\item{save}{Whether to save the image to file. (default: FALSE)}

\item{figwidth}{Width of the output image file in pixels (\code{px}).}

\item{figheight}{Height of the output image file in pixels (\code{px}).}

\item{out_path}{File path for the output image file. Existing files at this path are overwritten.
Whatever extension included in the image file name (e.g., jpg, png, pdf) will
be the format that is outputted.}
}
\value{
A \code{ggplot2} figure object. If save is \code{TRUE}, also exports the image to file.
}
\description{
This function creates a normalized stacked bar plot reporting frequencies of
a categorical variable grouped by a specified grouping variable.
}
