% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_styled_table_w_2_headers.R
\name{display_styled_table_w_2_headers}
\alias{display_styled_table_w_2_headers}
\title{Display a Styled HTML Table with Two Headers}
\usage{
display_styled_table_w_2_headers(data)
}
\arguments{
\item{data}{A data frame to be displayed in the styled HTML table.}
}
\value{
A styled HTML table object with two header rows.
}
\description{
This function takes a data frame and returns a style HTML table with two header rows.
}
