% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftOver_vcf.R
\name{liftover_vcf}
\alias{liftover_vcf}
\title{Liftover VCF Coordinates from GRCh37 to GRCh38}
\usage{
liftover_vcf(vcf, chain_path)
}
\arguments{
\item{vcf}{A \code{VCF} object (i.e., when reading in VCF via \code{VariantAnnotation::readVcf})
containing variant data including coordinates.}

\item{chain_path}{A character string specifying the file path to the chain file for conversion.}
}
\value{
A \code{VCF} object with updated coordinates. Only variants that successfully map
to a single location are retained.
}
\description{
This function lifts over variant coordinates in a VCF object using a chain
file.
}
\note{
The \code{seqlevel} of the VCF will be converted to \code{UCSC} in the mapping process.
}
