% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_barplot.R
\name{stratified_barplot}
\alias{stratified_barplot}
\title{Plot a stratified barplot with or without pairwise statistical testing}
\usage{
stratified_barplot(
  df,
  column,
  groups = NULL,
  ylab = NULL,
  xlab = NULL,
  legendlab = NULL,
  color_list = NULL,
  test = TRUE,
  multi_test_correct = NULL,
  alpha = 0.05,
  keep_caps = NULL,
  save = FALSE,
  figwidth = 1000,
  figheight = 1000,
  out_path = "stratified_barplot.jpg"
)
}
\arguments{
\item{df}{Dataframe containing the grouping variable and categorical variable of interest.}

\item{column}{Name of the categorical variable in \code{df} to be plotted.}

\item{groups}{Name of the categorical variable in \code{df} to group by. If not supplied, then
a non-stratified boxplot will be plotted.}

\item{ylab}{Title for the y-axis. (default is to use \verb{Frequency (\%)})}

\item{xlab}{Title for the x-axis. (default is to use name given to \code{groups})}

\item{legendlab}{Title for the legend. (default is to use name given to \code{column})}

\item{color_list}{Vector of R recognized color strings the length of the number of groups in
the variable provided to \code{column}.}

\item{test}{Whether to perform pairwise statistical testing. (default: \code{TRUE})}

\item{multi_test_correct}{Which method to use for multiple testing correction. Can specify any methods
that are accepted by the \code{p.adjust} function. (default is to perform no
correction)}

\item{alpha}{P-value threshold for significance. (default: 0.05)}

\item{keep_caps}{Vector of character strings to make sure to keep capitalized. The function
automatically tries to keep roman numerals capitalized, but any other string
needs to be provided here.}

\item{save}{Whether to save the image to file. (default: FALSE)}

\item{figwidth}{Width of the output image file in pixels (\code{px}).}

\item{figheight}{Height of the output image file in pixels (\code{px}).}

\item{out_path}{File path for the output image file. Existing files at this path are overwritten.
Whatever extension included in the image file name (e.g., jpg, png, pdf) will
be the format that is outputted.}
}
\value{
A \code{ggplot2} figure object. If save is \code{TRUE}, also exports the image to file.
}
\description{
This function creates a bar plot reporting frequencies of a categorical
variable grouped by a specified grouping variable. It can also test for
significant differences between groups using Fisher's exact test, annotating
significant comparisons in the plot.
}
