% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_violin_boxplot.R
\name{stratified_violin_boxplot}
\alias{stratified_violin_boxplot}
\title{Plot a Stratified Violin-Boxplot Using Plotly}
\usage{
stratified_violin_boxplot(
  data,
  var,
  strata,
  ylab = NULL,
  xlab = NULL,
  test = c("t.test", "wilcox.test"),
  alpha = 0.05
)
}
\arguments{
\item{data}{Dataframe containing the variables of interest.}

\item{var}{Name of the numerical variable in \code{data} to be plotted.}

\item{strata}{Name of the categorical variable in \code{data} to group by.}

\item{ylab}{Title for the y-axis (default is to use name given to \code{var}).}

\item{xlab}{Title for the x-axis (default is to use name given to \code{strata}).}

\item{test}{Which test to use for group comparison: "t.test" or "wilcox.test"
(default: "t.test").}

\item{alpha}{P-value threshold for significance (default: 0.05).}
}
\value{
A \code{plotly} figure object.
}
\description{
This function creates a violin-box plot with data points and means/stds for a
numeric variable grouped by a specified strata variable. It also can test for
significant differences between groups using a t-test or Wilcoxon rank-sum
test, annotating significant comparisons in the plot.
}
