% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_violin_boxplot.R
\name{stratified_violin_boxplot}
\alias{stratified_violin_boxplot}
\title{Plot a stratified violin-boxplot with or without pairwise statistical testing}
\usage{
stratified_violin_boxplot(
  df,
  column,
  groups = NULL,
  ylab = NULL,
  xlab = NULL,
  test = "fisher.test",
  multi_test_correct = NULL,
  alpha = 0.05,
  keep_caps = NULL,
  save = FALSE,
  figwidth = 1000,
  figheight = 1000,
  out_path = "stratified_violin_boxplot.jpg"
)
}
\arguments{
\item{df}{Dataframe containing the grouping variable and numeric variable of interest.}

\item{column}{Name of the numeric variable in \code{df} to be plotted.}

\item{groups}{Name of the categorical variable in \code{df} to group by. If not supplied, then
a non-stratified boxplot will be plotted.}

\item{ylab}{Title for the y-axis. (default is to use name given to \code{column})}

\item{xlab}{Title for the x-axis. (default is to use name given to \code{groups})}

\item{test}{Which test to use for group comparison: \code{t.test}, \code{wilcox.test}, or \code{NULL}
(no testing performed). (default: \code{wilcox.test})}

\item{multi_test_correct}{Which method to use for multiple testing correction. Can specify any methods
that are accepted by the \code{p.adjust} function. (default is to perform no
correction)}

\item{alpha}{P-value threshold for significance. (default: 0.05)}

\item{keep_caps}{Vector of character strings to make sure to keep capitalized. The function
automatically tries to keep roman numerals capitalized, but any other string
needs to be provided here.}

\item{save}{Whether to save the image to file. (default: FALSE)}

\item{figwidth}{Width of the output image file in pixels (\code{px}).}

\item{figheight}{Height of the output image file in pixels (\code{px}).}

\item{out_path}{File path for the output image file. Existing files at this path are overwritten.
Whatever extension included in the image file name (e.g., jpg, png, pdf) will
be the format that is outputted.}
}
\value{
A \code{ggplot2} figure object. If save is \code{TRUE}, also exports the image to file.
}
\description{
This function creates a violin-box plot with data points and means/stds for a
numeric variable grouped by a specified groups variable. It also can test for
significant differences between groups using a t-test or Wilcoxon rank-sum
test, annotating significant comparisons in the plot.
}
